renamed the comment route for a user getting their comments for better convention and tested it and removed the post id for easier set up in the front end fetching
